code/backend/requirements.txt

fastapi~=0.115.8
uvicorn~=0.34.0

pydantic~=2.9.2
requests~=2.31.0

sqlalchemy
mysql-connector-python

python-dotenv>=1.0.0
python-docx>=0.8.11
PyPDF2>=3.0.0
beautifulsoup4>=4.12.0
scikit-learn>=1.3.0

dashscope>=1.22.2
